clean up handling of canterp
LinuxCNC has support for "pluggable interpreters".  This feature lets a
user replace the default G-code interpreter with a different interpreter,
that understands a different input language but otherwise interacts with
the rest of the motion controller as usual.

LinuxCNC includes a pluggable interpreter module called "canterp",
short for "canonical interpreter".  This interpreter reads & parses
files containing canonical motion commands instead of G-code.

Before this commit we were mishandling the canterp module by treating
it like a shared library.  It was named libcanterp.so and was installed
in /usr/lib.

The canterp module is really more like our uspace realtime components:
it's a shared object file to be optionally and explicitly loaded into
an application with `dlopen()`.

This commit does a couple of things:

* renames canterp from `libcanterp.so` to just `canterp.so`
* moves canterp from `/usr/lib` to `/usr/lib/linuxcnc`
* updates the sample config to use the new name
* expands the testing of pluggable interpreters