interpmodule: fix `speed` property
This makes the `speed` property into a working array, instead of a
broken scalar.

The `interpreter.speed` property is intended to report the current
spindle speed.  It's been broken since the multi-spindle stuff went
in: the code compiled but failed at runtime, as demonstrated by the
`sim/axis/remap/getting-started/demo.ini` config.