halcompile: Add command line arguments to provide compile and link flags
It is possible to provide extra compiler and linker flags with "option" lines
inside a hal component written in the .comp preprocessor format.
However for files written in straight C this does not work.
This commit adds --extra-compile-args and --extra-link-args as command-line
options to halcompile.
See https://github.com/LinuxCNC/linuxcnc/issues/2204 for more background


Signed-off-by: andypugh <andy@bodgesoc.org>