force PyOpenGL to use GLX
On Bookworm and newer pyopengl detects if it's running on Wayland, and
chooses EGL there.  This is problematic because the rest of LinuxCNC is
still on GLX, and (as i understand it) one should not mix GLX and EGL
within a single application.

See <https://github.com/LinuxCNC/linuxcnc/issues/2264> for some details.

The correct fix is probably to teach our code to either use GLX everywhere
(for when running on X, and like it does after this commit) or EGL
everywhere (for when running on Wayland).

This commit fixes Axis on Wayland on Bookworm for me.