gmoccapy -add log levels to gmoccapy and getiniinfo
This will allow the user to control the many debug messages sent to the terminal.
It uses qtvcp's logging library so no reinventing the wheel.
Other files should probably be converted to use logging too.
Ultimately it would be nice if gmoccapy and qtvcp used the same INI parsing code too.