Fix "userspace" vs "realtime" nomenclature
Way back when, we had realtime happening in the kernel and non-realtime
happening in userspace.  We used the terms "realtime" and "kernel"
interchangably, and the terms "userspace" and "non-realtime"
interchangably.  In practice we usually referred to "realtime" and
"userspace" contexts, rather than "kernel" and "non-realtime" contexts,
but both forms of the terminology were used.

Now that we have Preempt-RT and uspace realtime, the old terminology is
inaccurate, misleading, and confusing to new users.

This commit updates the documentation around this to consistently refer to
"realtime" and "non-realtime" context, and avoid the old terms "kernel"
and "userspace" (where appropriate).