axis: copy gcode to temporary file before opening (fixes #2418)
The issue here is that if you open a gcode file, and then change the
file on disk, you'll think your changes are completely inert because
they don't show up in the GUI, but actually LinuxCNC has its own
separate file handle behind-the-scenes, and your machine can do scary
and unpredictable things.

With this fix, the code being executed by LinuxCNC always matches what
is displayed in the AXIS GUI even if you subsequently change the file.