motion: make & use a local copy of the emcmotCommand in the command handler
This does not change the behavior of Motion's command handler much.
It takes slightly longer to copy the entire command struct than it would
to access just the fields that the current command needs.

The race condition between Task and Motion accessing the command struct
is still there.  This commit lays the groundwork for protecting the
command struct and fixing the race, in the following commit.