mb2hal: wait for threads to stop before deallocating the memory they're using
This fixes a shutdown-time race condition in mb2hal between the main
thread and the per-link threads.

The main thread connects to all the Modbus interfaces and allocates all
HAL data structures, then starts a worker thread for each modbus link.

The main thread and the worker threads all watch a "time to shutdown"
status variable, which is set by a shutdown signal handler.  When the
"time to shutdown" variable gets set, the per-link threads all exit and
the main thread closes down the modbus communications and deallocates
the HAL memory.

But there was no synchronization between the shutdown of the main thread
and the shutdown of the worker threads, so sometimes the worker threads
were still accessing the stuff that the main thread was freeing, leading
to segfaults.

This commit makes the main thread join the worker threads before
deallocating anything.