interp: fix GE/LE operators to better handle double-precision comparison
This commit does four things:
1. Changes the LE/GE operators to better handle double-precision 
floating-point comparisons. When comparing two values as equal (or not 
equal) the EQ and NE operators considered the floating-point precision 
"issue" by comparing the difference between the two values to 
TOLERANCE_EQUAL, GE and LE did not.
2. Tidy the difference by using fabs()
3. Modify the TOLERANCE_EQUAL threshold to 1e-6.  Previously this was 
set to .0001 which is very close to a reasonable difference between two 
gcode coordinate values.
4. Update the documentation to reflect these change.