Use G38.3 in qt_auto_probe_tool.ngc to be able to check probe status.
There is a test after one of the G38.2 in question that look like this:

  O500 IF [#5070 EQ 0]
  G90
  O500 return [-3] ; indicate probe contact failure to epilog
  O500 endif

But the manual state that "If the probing operation failed, G38.2 and G38.4
will signal an error by posting an message on screen if the selected GUI
supports that. And by halting program execution.".  When I try to include a similar
test in my test program, the program is aborted and the test is never executed.
This make me believe the wrong G code is used.   Or is there something special
about the M6 remapping environment that make G38.2 behave differently?

Made sure both calls to G38.2 are changed to G38.2 with a test verifying the
status.

Also adjusted the handling of the measured height to take any coordinate system
offsets.