qtvcp -action_buttons: fix spindle up/dwn indicator bevaivour
If you changed the buttons from fwd/rev to up/dwn with style sheets
the checked state/indicator would be wrong.
Cause by a function callback that is called with fwd/red sction
that shouldn't be called with up/dwn action.
We can't easily disconnect the callback so we check and short circuit
it instead.

also slipped in that stop spindle button will be purposely enabled in MDI mode