debian: rules.in: export all necessary variables and pass path to src dir
After these changes, the dh_auto_XX commands will receive all
the necessary variables and parameters and pass the applicable
variables\parameters to the dh_XX commands. All dh commands
will be run in the src directory if they are not overridden.

That is, we need to write the override like this:
override_dh_auto_build-arch:
	dh_auto_build -- build-software

NOT so:
override_dh_auto_build-arch:
	dh_build -- build-software