override_dh_auto_clean: Perform cleaning only if required
The src/Makefile file has a check of variables at the very
beginning and exits by mistake if they are not initialized.
That is, if there is no src/Makefile.inc file, then the project
configuration was not performed and any other goals were not
fulfilled. It is not possible to complete any goal before
the initial configuration.
This check ensures that the source code is clean and there
is no need to clean it.