qtvcp -dro_widget:fix setting jontnumber with property.
This is a bit of a mess. the Qjoint_number property is actually axis
index, which can be out of sync with linuxcnc's joint number.
The 'is_homed' polish was using joint index rather the linuxcnc
joint number.
the joint type was only checked at initiation, so senting the joint number
 after linuxcnc started could have the wrong units type.

This needs to be thought thru some more to make this easier.
probably should use axis letter in axis mode and joint number in
joint mode.