spindle INI -fix non working M4 in some situations
Originally, if [SPINDLE_x] MAX_VELOCITY was set, the negative
max velocity was set to the same positive value, rather then
the negated speed.

Also changed the variable wording to help clearity

Also change MAX_VELOCITY to MAX_FWD_VELOCITY to be consistant and to
not clash with traditional PNCCONF (STEPCONF?) settings for stepgens.