Axis Limits check: Revert bad fix, add "re-check" This reverts https://github.com/LinuxCNC/linuxcnc/pull/1783 and attempts to fix https://github.com/LinuxCNC/linuxcnc/issues/2646
gcodemodule.cc accurately computes the joint limits, including tool offsets.
There is no need to add the last tool loaded on after the fact before testing
However, it only computes the extents at G-code load time
And the limits check is done when the Axis run button is pressed
If the tool lengths have changed (MDI, tool length probe...) then at the
point that the run button is pressed the g-code extents are out-of-date
This patch adds a new "re-check" button to the error dialog which reloads
the G-code than checks the limits again.
See issue 2646 for more