Improve treatment of o-codes in documentation
I recently learned the NGC coding language by reading the page currently called G-code overview. I found it unnecessarily confusing because all the remapping code I was trying to decipher uses o-codes and the documentation didn’t address them until after it had supposedly laid out all the acceptable ‘words.’ Also, in places the language describing them did not make much sense “1-30 subroutine local parameters of call arguments.” Additionally, sometimes o-code is used, and sometimes o-word, but neither was clearly defined nor the interchangeability of the terms addressed.

The reason for this seems benign: this document was started 20 years ago at NIST before the LinuxCNC community extended it to include o-codes.  When that happened, an o-code page was created, and some references were added into the lower parts of this overview document. But it seems attention to integrating an introduction to this functionality didn’t happen.

I also clarified rules on Comments, to move mention of “active” comments like (MSG) to the first mention of commenting.

I’m hopeful the community will accept my efforts to clarify for other newbs who find this page in the future.

I propose also changing the page title from “G-code Overview” to “NGC-code Overview” because it’s kind of a mis-nomer since the language is NGC, and also the page g-code.adoc is titled “G Codes” which is confusingly similar. I checked for another reference to the chapter heading, and did not find one, so I also changed the chapter heading from g-code-overview to ngc-code-overview.

On the O-code page, I added reference to the interchangeability of the terms o-code and o-word, and moved the “best practice” note to the introduction. I then changed all references to O-code or O-word to lower case o- to match that best practice.