qtvcp -qt_istat: look for _named_ INI MDI commands in the INI
The old way of just indexing the MDI_COMMAND = command list is not flexible enough.
Many objects assume they are the only thing reading the INI MDI COMMANDS.
Here we now look for named commands INI_COMMAND_sss, sss being any string.
In this way the order in the INI doesn't matter and objects can look for specific
named commands.

falls back to the legacy way for now, but puts out a warning