Be more tolerant to string in VERSION file
When the string in VERSION file contains date and git sha
like the following:
 2.10.0-0.96.20231128git2217582.fc38.src.rpm

then qtplasma refuses to work as it is unnable to properly
find 'major' and 'minor' components from the version string.

So let's parse the string from left where we always expect
'major' and 'minor' to be located.

Signed-off-by: Damian Wrobel <dwrobel@ertelnet.rybnik.pl>