Fixes more warnings for possible string truncation
This commit changes strncpy() for snprintf(), snprintf() guarantees a
NULL terminating string.

These places did not check the length at all, some hard-coded a NULL
terminating string as the last place in the array. With this, some of
the strings could potentially be without the NULL terminating character.

By changing to snprintf() we run into the possibility of truncating the
strings, no checks is currently in place to catch this.

There is also a strncat() thrown on here, hard-coded value was changed
from 1 to 2 to make room for the NULL terminating character.