Update Doxygen config file to version 1.9.4
This can be used for generating documentation from docstrings in the
source code. For now I believe we already have enough different types of
documentation but since this file is already present, it might as well
be up to date.

It requieres 'doxygen' to be available in PATH, to create HTML
documentation:
$ cd src/
$ doxygen