Fixes another address that can't be NULL
The variable 'ProcessName' is an array of 'char', it can't be NULL, so
don't check for it.

Make sure that the variable 'proc_name' is initialized to NULL. Change
how we copy the string from malloc() and strcpy() to strndup(). As with
malloc() we don't check the result, this is done before usage.