Gmoccapy, don't hide recurring error messages
The original code will suppress showing an error message if it is identical to the one shown last. While this seems to make some sense at first glance it leads to confusing situations where the user thinks an error has been fixed while it actually still exists but is hidden by the gui.