plasma M159: remove match syntax, improve error handling
On Python 3.7.3 (and maybe all <3.10), "match" is seen as syntax error, and therefore the script fails entirely.

Added more pointed error handling as a linuxcnc error so the operator gets a message, and because this affects a cut and could ruin material, the script aborts the running cylce if an error is encountered.