interp_g7x: Fix endless loop in recursive function
If the last path returns false in dive(), it's a high probability to end
up in an endless loop, there is one specific case that can prevent it.

This commit checks if we're on the last path and returns from the
recursive function.

The check is added before we calculate intersecting lines, since we're
on the last path, there shouldn't be any.