posemath: fix for arc to circle transition
fix for the issue discussed on the forum here:  https://forum.linuxcnc.org/plasmac/52584-inssues-hole-cutting

This particular issue did not affect imperial configs, only metric. This change gives more leeway on the floating-approximately-zero check as the fuzz values in posemath.h are the same regardless of the gui's units.  Commit 23bb2ec opened this up from 0.0, and this commit opens it up a bit further.