Dropped groff/nroff manual pages and migraded content to asciidoc.
The asciidoc manual pages are converted to groff during the build process.

The migration was done using scripts/man2adoc-migrate, with manual editing
to clean up the formatting, correct manual page redirects (files with only
groff .so statements), and adjusted file names and content to get this working.

Adjusted build rules to use new names for intro man pages. To avoid a name
conflict intro.3hal and intro.3rtapi was renamed to hal.3 and rtapi.3 (both
could not be generated to intro.3).