Adjusted handling of man pages to avoid duplicate lists.
Also adjusted d/configure to make it more obvious that the initial
settings for X are overwritten by later statements.

The list of manual pages to install are listed in linuxcnc.manpages.in,
there is no point in also listing them in linuxcnc.install.in.  Sorted
list in linuxcnc.manpages.in.

Partly based on patch from Steffen Moeller in https://github.com/LinuxCNC/linuxcnc/pull/2737/ .