Update per discussions
This update makes the following changes, per comments on 2/17/2024
1. Changes the stock remap to unconditionally return the machine to the position it was in when the remap was called. (This gets rid of a previous iteration that added a 'return_option' variable to the .ini file.)
2. Changes to documentation to reflect this change.

In full disclosure: This change will now also have the effect of largely reversing PR # 2562, because this version goes back to using G38.2. I do that because in practice G38.3 (the main change in PR 2562) does not allow the QtDragon implementation to display probing errors to the user. This this version will restore that functionality. This code does, however, retain the order of additional error-checking that was changed in PR # 2562.

These changes will also hopefully pass error checking because they are based on an updated fetch of 2.9.