tooledit_widget: decimal point = dot
Previously:
tooledit_widget displayed float numbers with decimal dot for localizations that use decimal dot
tooledit_widget displayed float numbers with decimal comma for localizations that use decimal comma

Now:
tooledit_widget displays float numbers with a decimal dot for all localizations

The functionality is preserved, when it is possible to enter a float with both a comma and a dot for localizations with a decimal comma

Reason for change:
Gmoccapy displayed some numbers with a decimal dot and some numbers with a decimal comma. This is a unification. More here:
https://github.com/LinuxCNC/linuxcnc/issues/2966#issuecomment-2254394542