gmoccapy/gladevcp/qtvcp -update Istat on instantiation
This returns to the original behaviour of setting all Istat data
immediately rather then waiting for update() to be called later.

It means it needs to be instatiated after the INI path has been
established. If no path argument is supplied, the environment will
be checked.

This fixes some scripts that assumed the environmental INI path
was used.