Increased single-step test timeout and include time limit in error message.
This test fail at random times when the build test is running from CI on github.
The failure seem to be a simple timeout, because the code has a limited amount
of seconds it will wait before giving up. Increased timeout from 5 to 15 seconds.

The change should reduce the chance of random failures during testing, and provide
some feedback to those reading the test log as to why it failed.