qtvcp calc: prevent display input, fix check for None
Currently inputting text directly into the display (5+5 for example) and clilcking = will cause a crash.  In the future perhaps this can be sorted out to accept this, for now it's best to prevent input into this box.

The newly added preference stores "None" as a string, fix the check to check for "None" instead of None.