qtplasmac: some more File Clear clean up from commit 40e4a23
With regard to running subroutine files via MDI calls, commit c9fe8ff fixed file-loaded being emitted for empty file names, however, in the case of QtPlasmaC, if the user had last clicked "File Clear", then hal_glib would trigger the file-loaded to be emitted because file_new and file_old did not match. QtPlasmaC was not ready for this (as this was not an issue previous to commit 40e4a23), and we never want to display the contents of "qtplasmac_file_clear.ngc".  This unnecessary file-loaded scenario is now it is accounted for.