qtplasmac updater: add helper function to expand ini file if #INCLUDE is used
Adds a helper function to attempt to make changes at the right location if a user decides to use #INCLUDE in their ini file as .expanded files are overwritten at run time.  This also helps when a user decides to try to update a sim config from a previous version rather than create a new sim config from the chooser.