gmoccapy/hal_bridge -allow hal_bridge to call macros in gmoccapy
This is a proof of concept to allow 3rd party (hal_bridge) to call
macros. The macros are run from Gmoccapy so no timing problems should
occur and and pre checks or post changes can be covered.

There needs to be agreement on where to put macro definitions in the INI.
Gmoccapy puts them under [MACROS], iniinfo under [DISPLAY]
python ZMQ module package must be available for this to work